fix(security): 0 attempts remaining, add back button
