+++ src/navigation/settings/SettingsNavigator.tsx
@@ -44,6 +44,7 @@ export type SettingsNavigationProp =
 	NativeStackNavigationProp<SettingsStackParamList>;
 
 export type SettingsStackParamList = {
+	AuthCheck: { onSuccess: () => void };
 	SettingsMenu: undefined;
 	GeneralSettings: undefined;
 	SecuritySettings: undefined;
@@ -67,7 +68,6 @@ export type SettingsStackParamList = {
 	BitcoinNetworkSelection: undefined;
 	LightningNodeInfo: undefined;
 	ManageSeedPhrase: undefined;
-	AuthCheck: undefined;
 	Channels: undefined;
 	ChannelDetails: undefined;
 	CloseConnection: undefined;
@@ -88,6 +88,7 @@ const SettingsNavigator = (): ReactElement => {
 	return (
 		<Stack.Navigator screenOptions={navOptions} initialRouteName="SettingsMenu">
 			<Stack.Group screenOptions={navOptions}>
+				<Stack.Screen name="AuthCheck" component={AuthCheck} />
 				<Stack.Screen name="SettingsMenu" component={SettingsMenu} />
 				<Stack.Screen name="GeneralSettings" component={GeneralSettings} />
 				<Stack.Screen name="SecuritySettings" component={SecuritySettings} />
@@ -130,7 +131,6 @@ const SettingsNavigator = (): ReactElement => {
 				/>
 				<Stack.Screen name="LightningNodeInfo" component={LightningNodeInfo} />
 				<Stack.Screen name="ManageSeedPhrase" component={ManageSeedPhrase} />
-				<Stack.Screen name="AuthCheck" component={AuthCheck} />
 				<Stack.Screen name="Channels" component={Channels} />
 				<Stack.Screen name="ChannelDetails" component={ChannelDetails} />
 				<Stack.Screen name="CloseConnection" component={CloseConnection} />
