+++ src/screens/Settings/Backup/ResetAndRestore.tsx
@@ -5,7 +5,6 @@ import { View, Text01S } from '../../../styles/components';
 import NavigationHeader from '../../../components/NavigationHeader';
 import Dialog from '../../../components/Dialog';
 import Button from '../../../components/Button';
-import AuthCheck from '../../../components/AuthCheck';
 import SafeAreaView from '../../../components/SafeAreaView';
 import Glow from '../../../components/Glow';
 import SafeAreaInsets from '../../../components/SafeAreaInsets';
@@ -28,55 +27,51 @@ const ResetAndRestore = ({
 					navigation.navigate('Tabs');
 				}}
 			/>
-			<AuthCheck>
-				<>
-					<View style={styles.container}>
-						<Text01S color="gray1">
-							Back up your wallet first to avoid loss of your funds and wallet
-							data. Resetting will overwrite your current Bitkit setup.
-						</Text01S>
+			<View style={styles.container}>
+				<Text01S color="gray1">
+					Back up your wallet first to avoid loss of your funds and wallet data.
+					Resetting will overwrite your current Bitkit setup.
+				</Text01S>
 
-						<View style={styles.imageContainer} pointerEvents="none">
-							<Glow style={styles.glow} size={600} color="brand" />
-							<Image source={imageSrc} style={styles.image} />
-						</View>
+				<View style={styles.imageContainer} pointerEvents="none">
+					<Glow style={styles.glow} size={600} color="brand" />
+					<Image source={imageSrc} style={styles.image} />
+				</View>
 
-						<View style={styles.buttonContainer}>
-							<Button
-								size="large"
-								variant="secondary"
-								style={styles.button}
-								text="Back Up First"
-								onPress={(): void => {
-									toggleView({
-										view: 'backupNavigation',
-										data: { isOpen: true },
-									});
-								}}
-							/>
-							<View style={styles.divider} />
-							<Button
-								size="large"
-								style={styles.button}
-								text="Reset Wallet"
-								onPress={(): void => setShowDialog(true)}
-							/>
-						</View>
-						<SafeAreaInsets type="bottom" />
-					</View>
-
-					<Dialog
-						visible={showDialog}
-						title="Reset Bitkit?"
-						description="Are you sure you want to reset your Bitkit Wallet? Do you have a backup of your recovery phrase and wallet data?"
-						onCancel={(): void => setShowDialog(false)}
-						onConfirm={(): void => {
-							wipeApp({});
-							setShowDialog(false);
+				<View style={styles.buttonContainer}>
+					<Button
+						size="large"
+						variant="secondary"
+						style={styles.button}
+						text="Back Up First"
+						onPress={(): void => {
+							toggleView({
+								view: 'backupNavigation',
+								data: { isOpen: true },
+							});
 						}}
 					/>
-				</>
-			</AuthCheck>
+					<View style={styles.divider} />
+					<Button
+						size="large"
+						style={styles.button}
+						text="Reset Wallet"
+						onPress={(): void => setShowDialog(true)}
+					/>
+				</View>
+				<SafeAreaInsets type="bottom" />
+			</View>
+
+			<Dialog
+				visible={showDialog}
+				title="Reset Bitkit?"
+				description="Are you sure you want to reset your Bitkit Wallet? Do you have a backup of your recovery phrase and wallet data?"
+				onCancel={(): void => setShowDialog(false)}
+				onConfirm={(): void => {
+					wipeApp({});
+					setShowDialog(false);
+				}}
+			/>
 		</SafeAreaView>
 	);
 };
